Delay Aggregation test and code with some bugs

+++ src/main/scala/sparktraining/model/CarrierDelayStats.scala
@@ -1,7 +1,11 @@
 package sparktraining.model
 
 case class CarrierDelayStats(carrier: String,
-                             minDelay: Long,
-                             maxDelay: Long,
-                             avgDelay: Long,
-                             medianDelay: Long)
+                             minDelay: Double,
+                             maxDelay: Double
+                             /*,
+                             avgDelay: Double,
+                             medianDelay: Double
+                             */
+                            )
+
